feat(utils): add pair type - [x] add pair type - [x] move csv/json to utils package
