+++ shared/utils/src/main/java/com/github/durex/utils/Pair.java
@@ -1,0 +1,13 @@
+package com.github.durex.utils;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor(staticName = "of")
+@NoArgsConstructor
+public class Pair<F, S> {
+  private F first;
+  private S second;
+}
