+++ shared/utils/src/test/java/com/github/durex/utils/JsonTest.java
@@ -1,4 +1,4 @@
-package com.github.durex.json;
+package com.github.durex.utils;
 
 import static com.github.durex.model.WeekDay.MONDAY;
 import static com.github.durex.pojo.MockUtils.getMockDirectory;
